// frontend/rodia-monorepo/apps/mobile/src/shared/api/generated/schemas/quoteStopResponse.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface QuoteStopResponse {
  quoteStopId?: number;
  seq?: number;
  address?: string;
  lat?: number;
  lng?: number;
  contactName?: string;
  contactPhone?: string;
  deptName?: string;
  managerName?: string;
}