// frontend/rodia-monorepo/apps/mobile/src/shared/api/generated/schemas/loadPlanResponse.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { CargoItem } from './cargoItem';
import type { Placement } from './placement';
import type { Stats } from './stats';

export interface LoadPlanResponse {
  placements?: Placement[];
  unplaced?: CargoItem[];
  stats?: Stats;
}