// frontend/rodia-monorepo/apps/mobile/src/shared/api/generated/schemas/truckSpecReferenceResponse.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface TruckSpecReferenceResponse {
  vehicleType?: string;
  vehicleTypeKr?: string;
  vehicleBodyType?: string;
  categoryKr?: string;
  tonnage?: number;
  maxWeight?: number;
  maxWeightDisplay?: string;
  maxVolume?: number;
  cargoLengthCm?: number;
  cargoWidthCm?: number;
  cargoHeightCm?: number;
  palletCount?: number;
  palletStandardMm?: string;
  doorPosition?: string;
  sourceName?: string;
}